Add a generic parser for Clearly Defined identifiers
Some dependencies cannot be expressed as part of the build system being
used. For example if a Maven based build requires a Mongo DB instance
running, then the POM would contain a dependency for the MongoDB client
but not for the DB itself. However, a Clearly Defined identifier for the
MongoDB source code location or archive could instead be fed into the
License Checker in such a case.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>