Add a notion of context.
This is a fairly large refactoring that introduces the notion of a
context to provide a central means of getting access to various
services. The intent is to provide a means of avoiding code duplication,
reducing complexity, and make more of the code more easily testable
(there's lots more work to do on testing specifically).