Fix #109 - ensure license check mojo is not inherited
The "inheritedByDefault" mojo annotation parameter is no longer used by
maven 3.8+ and the parameter was finally deprecated in maven-plugin-tools
version 3.6.2.

This causes repeat executions of the mojo in multi-module projects no
matter what we set "inheritedByDefault" to.

This change updates to latest maven-plugin-tools, removes usage of the
deprecated parameter and fixes the mojo to only run during execution of
the top level project in the reactor.

Also added an integration test.

Signed-off-by: Mat Booth <mat.booth@gmail.com>