Throw an exception to halt processing when an error is returned.
More robust exception handling is required, so this is stop gap (as is
most of the rest of the exception throwing). We need to make sure that
we stop processing when we encounter an error to avoid the case where a
failure results in the creation of needless and perhaps redundant review
requests.