Run license check on all events specified by the caller
This especially includes the 'push' event, so that the check can also
run on pushes to the master branch.

The use of events has been generalized with the intention to allow
callers to call the provided mavenLicenseCheck-workflow for any kind of
event they like and to treat only 'comment'-events in a special way.