Fix IP review links, so they are clickable from tool's output
Running `dash-licenses` in Theia's CI, to automatically create IP reviews
for suspicious dependencies, I encountered a small issue.

When such a dependency is found, looking at the log of the tool's run,
I tried clicking on the link to the review. However it failed because it
takes the end-of-sentence period, immediately following the URL, as being
part of it.

This PR adds a space between URL and the period, hopefully fixing the issue,
for both existing and new reviews.

Fixes #187

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>