[license] use EPL-2.0
The Theia project, from which the main script here is from originally,
has a secondary license (GPL-2.0-only WITH Classpath-exception-2.0).
It's useful there for distribution/integration of Theia by 3rd parties,
but would apparently serve no purpose for a build and test dependency
such as the dash-licenses-wrapper.

As of this writing, the original script, "check_3pp_licenses.js" [1] has
been contributed-to only by Ericsson employees, and so we have the
power to re-license if necessary (in the case where it's not proper to
drop the second part of the original license expression:
"OR GPL-2.0-only WITH Classpath-exception-2.0").

[1]: Paul Marechal, Vincent Fugnitto and myself:
   https://github.com/eclipse-theia/theia/blame/v1.43.0/scripts/check_3pp_licenses.js
Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>