docs(README): update the Gradle section
The bash version using grep is inconsistent, that's why an extensive
Gradle version, written in Groovy, is used to download the `dash.jar`
resolve all dependencies and run the Dash tool on them.

This allows for usage in ci/cd environments as the task will fail if the
Dash run fails.