[License-Workflow] Fail workflow on automated vetting-requests
Since https://github.com/eclipse/dash-licenses/pull/220 license-reviews
are automatically requested for PRs created by dependabot. But in that
change the logic of the 'Process license check results' step was not
updated to consider that case. This had the effect that the
license-vetting workflow always succeeded for PRs created by dependabot.

The intention to not fail the workflow in general when a license-review
was requested is because initially review-request were only created
manually from committers by adding a corresponding comment to the PR.
And that workflow execution should not fail respectively its result was
actually irrelevant.

Fixes https://github.com/eclipse-tycho/tycho/issues/3661